Adds testing
This uses the `apitest` library to do some nice "end-to-end" testing of the API.

This is not *yet* unit test coverage, but it is somewhere between E2E and unit tests. (Unit tests on APIs are awkward, because they involve mocking everything.)

In this case, the library lets me use the Gin router as my test surface, meaning I'm testing the authentic call. Because I implemented this with a CSP-style process network in the background, it slightly complicates things. But, the channel(s) provide a clean interface, meaning I am testing the process network, too.

My last test in the suite configures SQLite as an in-memory DB, and in doing so, lets me check that mocking a call to the API results in a value going all the way to SQLite, which I confirm by checking the in-memory DB.

So, in a way, nothing is mocked; everything is tested.