Adds summaries
This adds basic support for summarizing data at the start of a day (to process the data from yesterday).

There's more thought to go into this, but the summarization looks reasonable as a starting point.

Need to think about how to unit-test the summaries... probably need to have a deterministic database creation tool, and use that to then test that the summaries come out exactly as expected.