An ETL prototype
The goal would be a light pipeline spec, and the ability to expand it one element at a time.

It demonstrates that we could run summary calculations on the SQlite database. More work would be needed in order to (say) transfer data from an SQLite database to a destination Postgres DB.